pyproject.toml
[project]
name = "aivn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "google-auth>=2.48.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.65.0",
    "langfuse>=3.14.5",
    "moviepy>=2.2.1",
    "openinference-instrumentation-google-genai>=0.1.12",
    "pgvector>=0.4.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "rembg[cpu]>=2.0.69",
    "sqlalchemy>=2.0.47",
    "withoutbg>=1.0.2",
]
